# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ClaimInput": {
        "additionalProperties": false,
        "description": "Yêu cầu điều tra thô từ người dùng (API ``POST /api/v1/investigations``).",
        "properties": {
          "claim_text": {
            "maxLength": 5000,
            "minLength": 1,
            "title": "Claim Text",
            "type": "string"
          },
          "config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InvestigationConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "dose": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dose"
          },
          "drug": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Drug",
            "type": "string"
          },
          "event": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Event",
            "type": "string"
          },
          "population": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Population"
          },
          "route": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Route"
          },
          "time_window": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Window"
          }
        },
        "required": [
          "claim_text",
          "drug",
          "event"
        ],
        "title": "ClaimInput",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "InvestigationConfig": {
        "additionalProperties": false,
        "description": "Cấu hình nguồn và ngân sách runtime cho một cuộc điều tra.",
        "properties": {
          "max_documents": {
            "default": 50,
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Max Documents",
            "type": "integer"
          },
          "max_steps": {
            "default": 8,
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Max Steps",
            "type": "integer"
          },
          "sources": {
            "items": {
              "type": "string"
            },
            "title": "Sources",
            "type": "array"
          }
        },
        "title": "InvestigationConfig",
        "type": "object"
      },
      "LoginRequest": {
        "description": "Yêu cầu đăng nhập hoặc khởi tạo phiên.",
        "properties": {
          "password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password"
          },
          "role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Role"
              },
              {
                "type": "null"
              }
            ]
          },
          "token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          }
        },
        "title": "LoginRequest",
        "type": "object"
      },
...
      "Role": {
        "enum": [
          "investigator",
          "reviewer"
        ],
        "title": "Role",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "MvpApiToken": {
        "description": "Token theo vai trò (investigator/reviewer) gửi qua header X-API-Token.",
        "in": "header",
        "name": "X-API-Token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Nghiên cứu đối chiếu thuốc trên dữ liệu mô phỏng",
    "title": "MedReview VMEC-03",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "description": "Đăng nhập bằng token hoặc username/role và cấp cookie session HttpOnly.",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Login Api V1 Auth Login Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "description": "Đăng xuất, xóa session khỏi store và hủy cookie.",
        "operationId": "logout_api_v1_auth_logout_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "session_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Logout Api V1 Auth Logout Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Logout",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/investigations": {
      "get": {
        "description": "Danh sách cuộc điều tra đã lưu (mới nhất trước). Hỗ trợ lọc theo mine_only.",
        "operationId": "list_investigations_api_v1_investigations_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "mine_only",
            "required": false,
            "schema": {
              "default": false,
              "title": "Mine Only",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "MvpApiToken": []
          }
        ],
        "summary": "List Investigations",
        "tags": [
          "mvp"
        ]
      },
      "post": {
        "description": "Tạo cuộc điều tra mới và chạy nền; trả 202 + ID.",
        "operationId": "create_investigation_api_v1_investigations_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Investigation Api V1 Investigations Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "MvpApiToken": []
          }
        ],
        "summary": "Create Investigation",
        "tags": [
          "mvp"
        ]
      }
    },
...
  }
}